add components
