+++ src/components/templates/AccountImporter/account-importer.component.tsx
@@ -1,0 +1,38 @@
+import { Account } from "../../../types";
+import { useAccountImporter } from "./account-importer.hooks"
+
+export const AccountImporter = ({
+  accounts,
+  updateAccounts
+}: {
+  accounts: Account[], updateAccounts: (accounts: Account[]) => void
+}) => {
+  const {
+    isUploading,
+    csvData,
+    onUploadFile,
+    startImport,
+    onResetAccounts
+  } = useAccountImporter(accounts, updateAccounts);
+
+  return (
+    <div>
+      <div>
+          <p>Please choose csv</p>
+          <input type="file" onChange={(e) => onUploadFile(e)} />
+        </div>
+        <div>
+          <button disabled={isUploading || csvData === null} onClick={() => startImport()}>Import accounts</button>
+        </div>
+        <div>
+          <button onClick={() => onResetAccounts()}>reset accounts</button>
+        </div>
+        {/* print account count for each source */}
+        <div>
+          <p>icloud: {accounts.filter((account) => account.source === 'icloud').length}</p>
+          <p>chrome: {accounts.filter((account) => account.source === 'chrome').length}</p>
+          <p>firefox: {accounts.filter((account) => account.source === 'firefox').length}</p>
+        </div>
+    </div>
+  )
+}
